requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
dj-rest-auth-social==2.2.4
Django==4.1.3
django-allauth==0.50.0
django-cors-headers==3.13.0
django-filter==22.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==3.4
Markdown==3.4.1
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.0.10
pandas==1.5.1
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
